requirements.txt
SoundFile==0.10.3.post1
librosa==0.7.2
numpy==1.17.2
streamlit==0.72.0
Keras==2.4.3
split_folders==0.3.1
pandas==0.25.1
matplotlib==3.1.1
ipython==7.19.0
Pillow==8.1.0
scikit_learn==0.24.0
